test(community-export): move node-only PKC dataPath case out of node-and-browser
test.configs.pkc.test.ts imported getDefaultDataPath from the node runtime,
which fails to import on the browser leg and broke every browser CI job. Move
the dataPath-retention case to a new node-only file and drop the node-only
import (and now-unused PKCFactory) from the shared file.
